+++ doc/about/changes.md
@@ -4,7 +4,10 @@
 in progress
 {: .subtitle}
 
-No changes yet!
+### New features
+- Added `opts.static` to request static linking of libraries
+- MSVC's runtime can now be selected by using a combination of `opts.static`
+  and `opts.debug`
 
 ---